QcProc: Quality Control and Preprocessing

Selects ROI/AOI segments and target probes based on quality control and then generates gene-level count data.

QcProc(
  object,
  minSegmentReads = 1000,
  percentTrimmed = 80,
  percentStitched = 80,
  percentAligned = 80,
  percentSaturation = 50,
  minNegativeCount = 10,
  maxNTCCount = NULL,
  minNuclei = NULL,
  minArea = NULL,
  minProbeRatio = 0.1,
  outlierTestAlpha = 0.01,
  percentFailGrubbs = 20,
  removeLocalOutliers = TRUE
)

Arguments

object

object of class NanoStringGeoMxSet with raw data and metadata supplied

minSegmentReads

minimum number of reads

percentTrimmed

minimum % of reads trimmed

percentStitched

minimum % of reads stitched

percentAligned

minimum % of reads aligned

percentSaturation

minimum sequencing saturation

minNegativeCount

minimum negative control counts

maxNTCCount

maximum counts observed in NTC well

minNuclei

minimum # of nuclei estimated

minArea

minimum segment area

minProbeRatio

numeric between 0 and 1 to flag probes that have (geomean probe in all segments) / (geomean probes within target) less than or equal to this ratio

outlierTestAlpha

significance cut-off for Grubb's test to flag outlier probes

percentFailGrubbs

numeric to flag probes that fail Grubb's test in greater than or equal this percent of segments

removeLocalOutliers

boolean to determine if local outliers should be excluded from exprs matrix

Value

A list of two objects: QC-filtered NanoStringGeoMxSet object with gene-level targets as features and the QCFlags data frame appended to protocolData and QC plots

Details

A probe is removed globally from the dataset if either it does not reach the minProbeRatio threshold or the probe is an outlier according to Grubb’s test in at least percentFailGrubbs of segments. A probe is removed locally (from a given segment) if the probe is an outlier in that segment and removeLocalOtliers is set to TRUE (default).

The count for any gene with multiple probes per segment is calculated as the geometric mean of those probes.