Combine Nanostring Digital Spatial Profile read count and annotation files into a GeoMX object

StudyDesign returns a Sankey Plot and a GeoMX object

StudyDesign(
  dcc.files,
  pkc.files,
  pheno.data.file,
  pheno.data.sheet = "Template",
  pheno.data.dcc.col.name = "Sample_ID",
  protocol.data.col.names = c("aoi", "roi"),
  experiment.data.col.names = c("panel")
)

Arguments

dccFiles

A character vector containing the paths to the DCC files.

pkcFiles

A character vector representing the path to the corresponding PKC file

phenoDataFile

An optional character string representing the path to the corresponding phenotypic excel data file. It is recommended to use the Lab Worksheet in the exact order samples are provided in.

phenoDataSheet

An optional character string representing the excel sheet name containing the phenotypic data.

phenoDataDccColName

Character string identifying unique sample identifier column in phenoDataFile.

protocolDataColNames

Character list of column names from phenoDataFile containing data about the experimental protocol or sequencing data.

experimentDataColNames

Character list of column names from phenoDataFile containing data about the experiment's meta-data.

Value

A list containing the NanoString Object and the Sankey plot.

Details

For the function to run properly the annotation excel file must have the specific field names: 'slide name', class, segment, region

and the corresponding fields that can be renamed: Sample_ID, aoi, roi, panel