Run a linear mixed model on GeoMxSet

Run a linear mixed model on GeoMxSet

diffExpr(
  object,
  analysis.type,
  groups,
  group.col,
  regions,
  region.col,
  slide.col = "slide name",
  element = "q_norm",
  multi.core = TRUE,
  n.cores = 1,
  p.adjust = "BY",
  pairwise = NULL,
  fc.lim = 1.2,
  pval.lim.1 = 0.05,
  pval.lim.2 = 0.01
)

Arguments

object

Name of the NanoStringGeoMxSet to perform DE analysis on

analysis.type

Analysis type either "Between Groups" or "Within Groups"

groups

One or more groups of interest (needs to be variables within group.col)

group.col

Column for group (groups are usually found in different slides)

regions

One or more regions of interest (needs to be variables within region.col)

region.col

Column for region (regions are usually found within a slide)

slide.col

Column for slide name

element

assayDataElement of the geoMxSet object to run the DE on

multi.core

Set to TRUE to use multicore, FALSE to run in cluster mode (default is TRUE)

n.cores

Number of cores to use, set to 1 if running in serial mode (default is 1)

p.adjust

Method to use for pvalue adjustment. Choices are "holm","hochberg","hommel","bonferroni","BH","BY","fdr","none". Default is "BH"

pairwise

Boolean to calculate least-square means pairwise differences (default is NULL)

fc.lim

Fold Change limit for summarizing genes of interest (default is 1.2)

Value

a list containing mixed model output data frame, grid tables for samples and summary of genelists