Estimate cell composition across spatial samples using information from a custom cell signature matrix

Estimate cell composition across spatial samples using information from a custom cell signature matrix

spatialDeconvolution(
  dsp.obj,
  norm.expr.type = "",
  ref.mtx,
  ref.annot,
  cell.id = "CellID",
  celltype.col = "LabeledCellType",
  normalize = FALSE,
  min.cell.num = 0,
  min.genes = 10
)

Arguments

dsp.obj

Nanostring object containing normalized gene expression data

norm.expr.type

Name of the slot containing normalized gene expression data

ref.mtx

expression matrix (Gene x Reference_Samples)

ref.annot

annotated data.frame with cell.id and LabeledCellType, will be used along with ref.mtx to generate custom signature matrix

cell.id

column of annotated data.frame containing cell barcode or name

celltype.col

column of annotated data.frame containing cell identity information

normalize

scale profile matrix gene expression according to gene count

min.cell.num

minimum number of cells (within a type) required to generate signature matrix

min.genes

filters cells according to minimum number of genes expressed

Value

A list res containing the results of spatial deconvolution, res$beta: matrix of estimated cell abundances res$X: cell signature (gene x celltype) matrix used to generate deconvolution results res$yhat, res$resids: fitted values and log2-scale residuals from deconvolution, can be used to measure observations' goodness of fit res$prop_of_all: res$beta rescaled to proportion of celltype across each sample heatmap / barplot of estimated cell abundances in data