NanoString Digital Spatial Profile study

Exploring the normalized data by calculating the coefficient of variation (CV) for each gene (g) using the formula CVg=SDg/meang.

We then identify genes with high CVs that should have large differences across the various profiled segments.

This unbiased approach can reveal highly variable genes across the study.

We plot the results using unsupervised hierarchical clustering, displayed as a heatmap.

heatMap(
  target.data,
  norm.method = "quant",
  annotation.col = c("class", "segment", "region"),
  ngenes = 200,
  scale.by.row.or.col = "row",
  show.rownames = FALSE,
  show.colnames = FALSE,
  clustering.method = "average",
  cluster.rows = TRUE,
  cluster.cols = TRUE,
  clustering.distance.rows = "correlation",
  clustering.distance.cols = "correlation",
  annotation.row = NA,
  image.width = 3600,
  image.height = 1800,
  image.resolution = 300,
  image.filename = "heatmap.clust.highCVgenes.png",
  breaks.by.values = seq(-3, 3, 0.05),
  heatmap.color = colorRampPalette(c("blue", "white", "red"))(120)
)

Arguments

target.data

Normalized targetData S4 object input file.

norm.method

normalization method quant: Upper quartile (Q3) normalization and neg: background normalization

annotation.col

the annotations shown on right side of the heatmap

ngenes

Number of high CV genes to cluster and plot

scale.by.row.or.col

scale by row or by col ("row", "column" and "none")

show.rownames

boolean specifying if row names are be shown.

show.colnames

boolean specifying if column names are be shown.

clustering.method

clustering method used. Accepts the same values as hclust. e.g "ward.D", "ward.D2", "single", "complete", "average"

cluster.rows

boolean values determining if rows should be clustered or hclust object,

cluster.cols

boolean values determining if cols should be clustered or hclust object,

clustering.distance.rows

distance for clustering by rows (correlation, or euclidean)

clustering.distance.cols

distance for clustering by cols (correlation, or euclidean)

annotation.row

the annotations shown on left side of the heatmap

image.width

image width

image.height

image height

image.resolution

image resolution

image.filename

image filename

breaks.by.values

a sequence of numbers that covers the range of values in mat e.g seq(-3, 3, 0.05), 6/0.05=120 colors.

heatmap.color

colors of heatmap colorRampPalette(c("blue", "white", "red"))(120), here 120 is consistent with 120 above.

Value

A list containing the plot genes data matrix, and the heatmap plot.